Makefile.am: Install header files for library
The installs all library header files to ${prefix}/include/qsopt_ex.
The QSopt_ex.h header can be used to include all headers, or they
can be included separately. Note that the header files define a lot
of macros that are not namespace-prefixed and these can easily
interfere with the program they are included in.