README.md: Fix comment on license of original QSopt_ex
The website of the original QSopt_ex states that the license is
GPL3 or later. This was accidentally put as LGPL in the README.
The headers of the source files seem to simply state that GPL2 or
later applies. The License.txt file will be kept as GPL3 for now
since that is what seems to apply to the whole project.