contrib: Add initial Macports port file
The port file has a hard dependency on Python 2.7 and it installs
the Python module into the normal prefix. This means that the user
has to set PYTHONPATH to include this directory. This could be
resolved by using `PortGroup python` but this solution is tricky
since the PortGroup assumes that the Python module is in a package
that is independent of the library.