Add setup.py script to build Python module independently
The setup.py script is based on distutils. When building the module
using the autotools build system, the module will depend on the
simultaneously built libqsopt_ex. This build script can be used
to build and install the Python module based on the libqsopt_ex
that is already installed on the system. This makes building and
packaging easier in some cases, particularly when packaging for
various distributions.