Move library files in src/ to new directory qsopt_ex/
This means that header files can be referenced in the same way as
when in installed, using `<qsopt_ex/XX.h>`. This means that programs
and the Python module can more easily be built separately.

The esolver program file is split off into esolver/ and the test
programs eg_sloan and demo_qs are split off into tests/. Template
test programs are also put in tests/ but are not set up to build
yet in Makefile.am.