Explicitly use symbols with typename macro prefix
This commit puts the typename macro prefix EGLPNUM_TYPENAME on every
use of a global variable, macro or function that is specialized for
a specific type using the template system. In addition the actual
type uses the macro EGLPNUM_TYPE (previously EGlpNum_t).

This means that ctags is no longer needed to discover the names of
the public symbols in each header files, and as a result the
template substitution is much faster. Also, the generated files
no longer depends on the central templatetags file so changing one
header file will no longer cause all other template files to be
regenerated.